# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Les Canetons API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "https://lescanetons.org/api/v1",
            "description": "Production"
        }
    ],
...
    "paths": {
...
        "/me/password": {
            "post": {
                "operationId": "account.password",
                "description": "Any logged-in account holder, for their own account. No permission is\nneeded: this is the screen every member has and nobody administers, and\nit is where a member sent by `mustChangePassword` lands.\n\nSend `currentPassword` and `newPassword`. The new one must be at least\neight characters.\n\nOn success the password is replaced, the forced-change flag is cleared,\nand every other session the member has open is ended. Answers\n`{\"ok\": true, \"sessionsEnded\": n}`, where `sessionsEnded` counts those\nother sessions; the calling session stays valid.\n\nA wrong `currentPassword` answers `403 reauth_failed`. After five wrong\nones the account answers `429 too_many_attempts` for fifteen minutes,\nand a correct password during that window is still refused. A\n`newPassword` shorter than eight characters answers\n`400 validation_failed` with `too_short` against `newPassword`.",
                "summary": "Change your own password",
                "tags": [
                    "Session"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "sessionsEnded": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "sessionsEnded"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountPasswordRequest": {
                "type": "object",
                "description": "The credentials for changing your own password.\n\nSend the password the account is using now and the one to replace it with.\nThe current password is checked again server-side before anything is\nwritten, so knowing it is what separates a borrowed session from a\npermanently stolen account.\n\nThe only requirement on the new password is a minimum of eight characters.\nThere is no complexity rule, no reuse check and no expiry.",
                "properties": {
                    "currentPassword": {
                        "type": "string",
                        "description": "The password the account is using right now. Re-verified before the change is applied; a wrong one answers `403 reauth_failed`."
                    },
                    "newPassword": {
                        "type": "string",
                        "description": "The replacement. At least 8 characters; a shorter one fails with `too_short`.",
                        "minLength": 8,
                        "maxLength": 255
                    }
                },
                "required": [
                    "currentPassword",
                    "newPassword"
                ],
                "title": "AccountPasswordRequest"
            },
...
        },
        "responses": {
...
            "ValidationException": {
                "description": "The submitted fields were rejected. `errors` names each one and why.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "description": "English message. Never displayed: the front end renders `code`."
                                },
                                "status": {
                                    "type": "integer",
                                    "description": "The HTTP status code, repeated in the body as RFC 9457 intends."
                                },
                                "instance": {
                                    "type": "string",
                                    "description": "The path that was requested. Never the query string."
                                },
                                "code": {
                                    "type": "string",
                                    "description": "Stable machine token the front end maps to French.",
                                    "enum": [
                                        "validation_failed"
                                    ]
                                },
                                "errors": {
                                    "type": "array",
                                    "description": "One entry per rejected field, first failure only.",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "description": "The camelCase name of the rejected field, as it was submitted."
                                            },
                                            "reason": {
                                                "type": "string",
                                                "description": "Stable machine token saying what is wrong with that field.",
                                                "enum": [
                                                    "required",
                                                    "too_long",
                                                    "too_short",
                                                    "invalid_format",
                                                    "invalid_type",
                                                    "invalid_value",
                                                    "invalid_number",
                                                    "already_taken",
                                                    "must_be_after"
                                                ]
                                            },
                                            "params": {
                                                "type": "object",
                                                "additionalProperties": {}
                                            }
                                        },
                                        "required": [
                                            "field",
                                            "reason"
                                        ]
                                    }
                                },
                                "requestId": {
                                    "type": "string",
                                    "description": "ULID identifying this request. Echoed as the X-Request-Id header, and in the logs."
                                },
                                "detail": {
                                    "type": "string",
                                    "description": "RFC 9457 \u00a73.1.1: what happened and what to do about it, in English, for a developer reading the response. Never rendered to an end user."
                                }
                            },
                            "required": [
                                "title",
                                "status",
                                "code",
                                "instance",
                                "errors",
                                "requestId",
                                "detail"
                            ]
                        }
                    }
                }
            },
            "Problem419": {
                "description": "The CSRF token was missing or stale. Re-prime it and retry; you are still logged in.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "description": "English message. Never displayed: the front end renders `code`."
                                },
                                "status": {
                                    "type": "integer",
                                    "description": "The HTTP status code, repeated in the body as RFC 9457 intends."
                                },
                                "instance": {
                                    "type": "string",
                                    "description": "The path that was requested. Never the query string."
                                },
                                "code": {
                                    "type": "string",
                                    "description": "Stable machine token the front end maps to French.",
                                    "enum": [
                                        "invalid_session"
                                    ]
                                },
                                "errors": {
                                    "type": "array",
                                    "description": "Always empty for this status; present so every problem has the same shape.",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "requestId": {
                                    "type": "string",
                                    "description": "ULID identifying this request. Echoed as the X-Request-Id header, and in the logs."
                                },
                                "detail": {
                                    "type": "string",
                                    "description": "RFC 9457 \u00a73.1.1: what happened and what to do about it, in English, for a developer reading the response. Never rendered to an end user."
                                }
                            },
                            "required": [
                                "title",
                                "status",
                                "code",
                                "instance",
                                "errors",
                                "requestId",
                                "detail"
                            ]
                        }
                    }
                }
            },
...
            "AuthenticationException": {
                "description": "No session, or it has expired. Log in and retry.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "description": "English message. Never displayed: the front end renders `code`."
                                },
                                "status": {
                                    "type": "integer",
                                    "description": "The HTTP status code, repeated in the body as RFC 9457 intends."
                                },
                                "instance": {
                                    "type": "string",
                                    "description": "The path that was requested. Never the query string."
                                },
                                "code": {
                                    "type": "string",
                                    "description": "Stable machine token the front end maps to French.",
                                    "enum": [
                                        "not_authenticated"
                                    ]
                                },
                                "errors": {
                                    "type": "array",
                                    "description": "Always empty for this status; present so every problem has the same shape.",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "requestId": {
                                    "type": "string",
                                    "description": "ULID identifying this request. Echoed as the X-Request-Id header, and in the logs."
                                },
                                "detail": {
                                    "type": "string",
                                    "description": "RFC 9457 \u00a73.1.1: what happened and what to do about it, in English, for a developer reading the response. Never rendered to an end user."
                                }
                            },
                            "required": [
                                "title",
                                "status",
                                "code",
                                "instance",
                                "errors",
                                "requestId",
                                "detail"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}